+++ cmd/hidden_lake/helpers/encryptor/pkg/settings/settings.go
@@ -6,6 +6,7 @@ const (
 )
 
 const (
+	CPathKey = "hle.key"
 	CPathYML = "hle.yml"
 )