+++ cmd/hidden_lake/service/pkg/settings/settings.go
@@ -15,6 +15,7 @@ const (
 )
 
 const (
+	CPathKey = "hls.key"
 	CPathYML = "hls.yml"
 	CPathDB  = "hls.db"
 )